pyproject.toml
[project]
name = "mlx-whisper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "datasets>=4.0.0",
    "dotenv>=0.9.9",
    "ipykernel>=6.29.5",
    "matplotlib>=3.9.4",
    "openai-whisper>=20250625",
    "pandas>=2.3.1",
    "soundfile>=0.13.1",
    "streamlit>=1.46.1",
    "streamlit-webrtc>=0.63.3",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchcodec>=0.4.0",
    "translate>=3.6.1",
    "wandb>=0.21.0",
]
